Move FtdiDevice and FtdiCore from 'gui' package to 'common'
